GreenDao优化

+++ app/src/main/java/com/example/greendaodemo/db/base/BaseDao.java
@@ -6,6 +6,7 @@ import android.database.sqlite.SQLiteConstraintException;
 import com.example.greendaodemo.db.DaoThreadPool;
 
 import org.greenrobot.greendao.AbstractDao;
+import org.greenrobot.greendao.database.StandardDatabase;
 import org.greenrobot.greendao.query.QueryBuilder;
 
 import java.util.List;
@@ -20,7 +21,7 @@ public class BaseDao<T, K> {
     private Activity activity;//传activity就在主线程回调，不传就在子线程回调
 
     public BaseDao(AbstractDao dao) {
-        mDao = dao;
+        this.mDao = dao;
     }
 
     public BaseDao(AbstractDao dao, Activity activity) {
@@ -29,17 +30,22 @@ public class BaseDao<T, K> {
     }
 
     //将给定的实体插入数据库
-    public long save(T item) {
+    public void save(T item) {
+        if (isDBClose()) {
+            return;
+        }
         try {
-            return mDao.insert(item);
+            mDao.insert(item);
         } catch (SQLiteConstraintException e) {
             e.printStackTrace();
         }
-        return 0;
     }
 
     //使用事务操作，将给定的实体插入数据库
     public void save(T... items) {
+        if (isDBClose()) {
+            return;
+        }
         try {
             mDao.insertInTx(items);
         } catch (SQLiteConstraintException e) {
@@ -49,6 +55,9 @@ public class BaseDao<T, K> {
 
     //使用事务操作，将给定的实体集合插入数据库
     public void save(List<T> items) {
+        if (isDBClose()) {
+            return;
+        }
         try {
             mDao.insertInTx(items);
         } catch (SQLiteConstraintException e) {
@@ -57,93 +66,144 @@ public class BaseDao<T, K> {
     }
 
     //将给定的实体插入数据库，若此实体类存在，则覆盖
-    public long saveOrUpdate(T item) {
-        return mDao.insertOrReplace(item);
+    public void saveOrUpdate(T item) {
+        if (isDBClose()) {
+            return;
+        }
+        mDao.insertOrReplace(item);
     }
 
     //使用事务操作，将给定的实体插入数据库，若此实体类存在，则覆盖
     public void saveOrUpdate(T... items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.insertOrReplaceInTx(items);
     }
 
     //使用事务操作，将给定的实体集合插入数据库，若此实体类存在，则覆盖
     public void saveOrUpdate(List<T> items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.insertOrReplaceInTx(items);
     }
 
     //从数据库中删除给定Key所对应的实体
     public void deleteByKey(K key) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.deleteByKey(key);
     }
 
     //从数据库中删除给定的实体
     public void delete(T item) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.delete(item);
     }
 
     //使用事务操作删除数据库中给定的实体
     public void delete(T... items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.deleteInTx(items);
     }
 
     //使用事务操作删除数据库中给定实体集合中的实体
     public void delete(List<T> items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.deleteInTx(items);
     }
 
     //删除数据库中全部数据
     public void deleteAll() {
+        if (isDBClose()) {
+            return;
+        }
         mDao.deleteAll();
     }
 
     //更新给定的实体
     public void update(T item) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.update(item);
     }
 
     //使用事务操作，更新给定的实体
     public void update(T... items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.updateInTx(items);
     }
 
     //使用事务操作更新数据库中给定实体集合中的实体
     public void update(List<T> items) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.updateInTx(items);
     }
 
     //加载给定主键的实体
     public T query(K key) {
+        if (isDBClose()) {
+            return null;
+        }
         return mDao.load(key);
     }
 
     //加载数据库中所有的实体
     public List<T> queryAll() {
+        if (isDBClose()) {
+            return null;
+        }
         return mDao.loadAll();
     }
 
     //根据条件查询集合
     public List<T> query(String where, String... params) {
+        if (isDBClose()) {
+            return null;
+        }
         return mDao.queryRaw(where, params);
     }
 
     //返回queryBuilder
-    public QueryBuilder<T> queryBuilder() {
-        return mDao.queryBuilder();
-    }
+//    public QueryBuilder<T> queryBuilder() {
+//        return mDao.queryBuilder();
+//    }
 
     //获取数据库中数据的数量
     public long count() {
+        if (isDBClose()) {
+            return 0;
+        }
         return mDao.count();
     }
 
     //刷新缓存
     public void refresh(T item) {
+        if (isDBClose()) {
+            return;
+        }
         mDao.refresh(item);
     }
 
     //清除缓存
-    public boolean detach(T item) {
-        return mDao.detach(item);
+    public void detach(T item) {
+        if (isDBClose()) {
+            return;
+        }
+        mDao.detach(item);
     }
 
     //将给定的实体插入数据库
@@ -151,6 +211,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 try {
                     final long insert = mDao.insert(item);
                     runUI(new Runnable() {
@@ -178,6 +241,7 @@ public class BaseDao<T, K> {
                 }
             }
         });
+
     }
 
     //使用事务操作，将给定的实体插入数据库
@@ -185,6 +249,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 try {
                     mDao.insertInTx(items);
                     runUI(new Runnable() {
@@ -208,6 +275,7 @@ public class BaseDao<T, K> {
                 }
             }
         });
+
     }
 
     //使用事务操作，将给定的实体集合插入数据库
@@ -215,6 +283,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 try {
                     mDao.insertInTx(items);
                     runUI(new Runnable() {
@@ -238,6 +309,7 @@ public class BaseDao<T, K> {
                 }
             }
         });
+
     }
 
     //将给定的实体插入数据库，若此实体类存在，则覆盖
@@ -245,6 +317,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final long insertOrReplace = mDao.insertOrReplace(item);
                 //UI
@@ -265,6 +340,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.insertOrReplaceInTx(items);
                 //UI
@@ -285,6 +363,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.insertOrReplaceInTx(items);
                 //UI
@@ -305,6 +386,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.deleteByKey(key);
                 //UI
@@ -325,6 +409,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.delete(item);
                 //UI
@@ -345,6 +432,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.deleteInTx(items);
                 //UI
@@ -365,6 +455,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.deleteInTx(items);
                 //UI
@@ -385,6 +478,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.deleteAll();
                 //UI
@@ -405,6 +501,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.update(item);
                 //UI
@@ -425,6 +524,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.updateInTx(items);
                 //UI
@@ -445,6 +547,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.updateInTx(items);
                 //UI
@@ -465,6 +570,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final T load = mDao.load(key);
                 //UI
@@ -485,6 +593,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final List<T> loadAll = mDao.loadAll();
                 //UI
@@ -505,6 +616,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final List<T> queryRaw = mDao.queryRaw(where, params);
                 //UI
@@ -520,11 +634,22 @@ public class BaseDao<T, K> {
         });
     }
 
+    //返回queryBuilder
+    public QueryBuilder<T> queryBuilder() {
+        if (isDBClose()) {
+            return null;
+        }
+        return mDao.queryBuilder();
+    }
+
     //获取数据库中数据的数量
     public void count(final OnDaoExcuteCallBack<Long> onDaoExcuteCallBack) {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final long count = mDao.count();
                 //UI
@@ -545,6 +670,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 mDao.refresh(item);
                 //UI
@@ -565,6 +693,9 @@ public class BaseDao<T, K> {
         executeSingleRunnable(new Runnable() {
             @Override
             public void run() {
+                if (isDBClose()) {
+                    return;
+                }
                 //数据库操作
                 final boolean detach = mDao.detach(item);
                 //UI
@@ -582,6 +713,7 @@ public class BaseDao<T, K> {
                 });
             }
         });
+
     }
 
     private void runUI(Runnable runnable) {
@@ -598,5 +730,10 @@ public class BaseDao<T, K> {
 
     public interface OnDaoExcuteCallBack<Bean> {
         void OnDaoExcuteResult(Throwable throwable, Bean bean);//throwable==null代表成功
+    }
+
+    //数据库是否关闭
+    public boolean isDBClose() {
+        return !((StandardDatabase) mDao.getDatabase()).getSQLiteDatabase().isOpen();
     }
 }
